♻️ move schemas
